Правки с api
